🚚 Move TopAppBar component to `appbar` package
